benovate_test/requirements.txt
django
djangorestframework
django-webpack-loader